experiments/test_agent/test_mcp/requirements.txt
fastmcp>=0.1.0
mcp>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
